Add main sfs and bind mount support
Arm builds do not make an iso, but can optionally make a main sfs.

Symlinks cannot leave a chroot but bind mounts allow
local-repositories to be placed on a different filesystem in case
SHARE is a mounted network block device (/dev/nbd0) and you want
to keep local-repositories on a local filesystem.